+++ src/AltinnCore/Designer/Controllers/ConfigController.cs
@@ -1,7 +1,16 @@
-﻿using AltinnCore.Common.Services.Interfaces;
+﻿using System.IO;
+using System.Text;
+using AltinnCore.Common.Configuration;
+using AltinnCore.Common.Helpers;
+using AltinnCore.Common.Services.Interfaces;
+using AltinnCore.ServiceLibrary.Configuration;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Hosting;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Options;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Linq;
 
 namespace AltinnCore.Designer.Controllers
 {
@@ -13,16 +22,22 @@ namespace AltinnCore.Designer.Controllers
     {
         private readonly IHostingEnvironment _hostingEnvironment;
         private readonly IRepository _repository;
+        private readonly ServiceRepositorySettings _settings;
+        private readonly IHttpContextAccessor _httpContextAccessor;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="ConfigController"/> class
         /// </summary>
         /// <param name="hostingEnvironment">The hosting environment service</param>
         /// <param name="serviceRepositoryService">The serviceRepository service</param>
-        public ConfigController(IHostingEnvironment hostingEnvironment, IRepository serviceRepositoryService)
+        /// <param name="repositorySettings">The repository settings</param>
+        /// <param name="httpContextAccessor">The http context accessor</param>
+        public ConfigController(IHostingEnvironment hostingEnvironment, IRepository serviceRepositoryService, IOptions<ServiceRepositorySettings> repositorySettings, IHttpContextAccessor httpContextAccessor)
         {
             _hostingEnvironment = hostingEnvironment;
             _repository = serviceRepositoryService;
+            _settings = repositorySettings.Value;
+            _httpContextAccessor = httpContextAccessor;
         }
 
         /// <summary>
@@ -97,6 +112,65 @@ namespace AltinnCore.Designer.Controllers
             JsonResult result = new JsonResult(configJson);
 
             return result;
+        }
+
+        /// <summary>
+        /// Method to retrieve the service description from the metadata file
+        /// </summary>
+        /// <param name="owner">the owner of the service</param>
+        /// <param name="service">the service</param>
+        /// <returns>The service description of the service</returns>
+        [HttpGet]
+        public string GetServiceDescription(string owner, string service)
+        {
+            string serviceMetadataDirectoryPath = _settings.GetMetadataPath(owner, service, AuthenticationHelper.GetDeveloperUserName(_httpContextAccessor.HttpContext)) + _settings.GetMetadataJsonFile();
+            string serviceDescription = string.Empty;
+
+            if (System.IO.File.Exists(serviceMetadataDirectoryPath))
+            {
+                string serviceConfiguration = System.IO.File.ReadAllText(serviceMetadataDirectoryPath, Encoding.UTF8);
+                ServiceConfiguration serviceConfigurationObject = JsonConvert.DeserializeObject<ServiceConfiguration>(serviceConfiguration);
+                if (serviceConfigurationObject != null)
+                {
+                    serviceDescription = serviceConfigurationObject.ServiceDescrition;
+                }
+            }
+
+            return serviceDescription;
+        }
+
+        /// <summary>
+        /// Method to set the service description in the metadata file
+        /// </summary>
+        /// <param name="owner">the owner of the service</param>
+        /// <param name="service">the service</param>
+        /// <param name="description">the service description</param>
+        [HttpPost]
+        public void SetServiceDescription(string owner, string service, [FromBody] dynamic description)
+        {
+            string serviceMetadataDirectoryPath = _settings.GetMetadataPath(owner, service, AuthenticationHelper.GetDeveloperUserName(_httpContextAccessor.HttpContext)) + _settings.GetMetadataJsonFile();
+            ServiceConfiguration serviceConfigurationObject = null;
+
+            if (System.IO.File.Exists(serviceMetadataDirectoryPath))
+            {
+                string serviceConfiguration = System.IO.File.ReadAllText(serviceMetadataDirectoryPath, Encoding.UTF8);
+                serviceConfigurationObject = JsonConvert.DeserializeObject<ServiceConfiguration>(serviceConfiguration);
+                serviceConfigurationObject.ServiceDescrition = description.serviceDescription.ToString();
+            }
+            else
+            {
+                new FileInfo(serviceMetadataDirectoryPath).Directory.Create();
+                serviceConfigurationObject = new ServiceConfiguration()
+                {
+                    RepositoryName = service,
+                    ServiceDescrition = description.serviceDescription.ToString(),
+                };
+            }
+
+            if (serviceConfigurationObject != null)
+            {
+                System.IO.File.WriteAllText(serviceMetadataDirectoryPath, JObject.FromObject(serviceConfigurationObject).ToString(), Encoding.UTF8);
+            }
         }
     }
 }
